+++ one/params.py
@@ -110,14 +110,18 @@ def setup(client=None, silent=False, make_default=None, username=None, cache_dir
     """
     # First get default parameters
     par_default = default()
-    client_key = _key_from_url(client or par_default.ALYX_URL)
+    default_url = par_default.ALYX_URL
+    client_key = _key_from_url(client or default_url)
 
     # If a client URL has been provided, set it as the default URL
-    par_default = par_default.set('ALYX_URL', client or par_default.ALYX_URL)
-    
-    # When silent=True, use default parameters instead of current ones to reset credentials
-    par_current = (par_default if silent else 
-                  iopar.read(f'{_PAR_ID_STR}/{client_key}', par_default))
+    par_default = par_default.set('ALYX_URL', client or default_url)
+
+    # When silent=True, if setting up default database use default parameters
+    # instead of current ones to reset credentials
+    if silent and client_key == _key_from_url(default_url):
+        par_current = par_default
+    else:
+        par_current = iopar.read(f'{_PAR_ID_STR}/{client_key}', par_default)
     if username:
         par_current = par_current.set('ALYX_LOGIN', username)